+++ src/ext-test/java/org/opentripplanner/ext/geocoder/LuceneIndexTest.java
@@ -4,7 +4,11 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.opentripplanner.transit.model._data.TransitModelForTest.id;
 import static org.opentripplanner.transit.model._data.TransitModelForTest.station;
 import static org.opentripplanner.transit.model._data.TransitModelForTest.stop;
+import static org.opentripplanner.transit.model.basic.TransitMode.BUS;
+import static org.opentripplanner.transit.model.basic.TransitMode.FERRY;
 
+import com.google.common.collect.ImmutableMultimap;
+import com.google.common.collect.Multimap;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -14,9 +18,11 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.opentripplanner.routing.graph.Graph;
+import org.opentripplanner.transit.model.basic.TransitMode;
 import org.opentripplanner.transit.model.framework.Deduplicator;
 import org.opentripplanner.transit.model.site.RegularStop;
 import org.opentripplanner.transit.model.site.Station;
+import org.opentripplanner.transit.model.site.StopLocation;
 import org.opentripplanner.transit.service.DefaultTransitService;
 import org.opentripplanner.transit.service.StopModel;
 import org.opentripplanner.transit.service.TransitModel;
@@ -35,11 +41,13 @@ class LuceneIndexTest {
 
   static RegularStop ALEXANDERPLATZ_BUS = stop("Alexanderplatz Bus")
     .withCoordinate(52.52277, 13.41046)
+    .withVehicleType(BUS)
     .withParentStation(ALEXANDERPLATZ_STATION)
     .build();
 
   static RegularStop ALEXANDERPLATZ_RAIL = stop("Alexanderplatz S-Bahn")
     .withCoordinate(52.52157, 13.41123)
+    .withVehicleType(TransitMode.RAIL)
     .withParentStation(ALEXANDERPLATZ_STATION)
     .build();
   static RegularStop LICHTERFELDE_OST_1 = stop("Lichterfelde Ost")
@@ -50,18 +58,27 @@ class LuceneIndexTest {
     .withId(id("lichterfelde-gleis-2"))
     .withCoordinate(52.42985, 13.32807)
     .build();
+  static RegularStop WESTHAFEN = stop("Westhafen")
+    .withVehicleType(null)
+    .withCoordinate(52.42985, 13.32807)
+    .build();
 
   // Atlanta
   static Station FIVE_POINTS_STATION = station("Five Points")
     .withCoordinate(33.753899, -84.39156)
     .build();
 
-  static RegularStop ARTS_CENTER = stop("Arts Center").withCoordinate(52.52277, 13.41046).build();
+  static RegularStop ARTS_CENTER = stop("Arts Center")
+    .withCode("4456")
+    .withCoordinate(52.52277, 13.41046)
+    .build();
   static RegularStop ARTHUR = stop("Arthur Langford Jr Pl SW at 220")
     .withCoordinate(52.52277, 13.41046)
     .build();
 
   static LuceneIndex index;
+
+  static StopClusterMapper mapper;
 
   @BeforeAll
   static void setup() {
@@ -72,6 +89,7 @@ class LuceneIndexTest {
         ALEXANDERPLATZ_RAIL,
         LICHTERFELDE_OST_1,
         LICHTERFELDE_OST_2,
+        WESTHAFEN,
         ARTS_CENTER,
         ARTHUR
       )
@@ -80,8 +98,23 @@ class LuceneIndexTest {
       .of(ALEXANDERPLATZ_STATION, BERLIN_HAUPTBAHNHOF_STATION, FIVE_POINTS_STATION)
       .forEach(stopModel::withStation);
     var transitModel = new TransitModel(stopModel.build(), new Deduplicator());
-    var transitService = new DefaultTransitService(transitModel);
+    var transitService = new DefaultTransitService(transitModel) {
+      private final Multimap<StopLocation, TransitMode> modes = ImmutableMultimap
+        .<StopLocation, TransitMode>builder()
+        .putAll(WESTHAFEN, FERRY, BUS)
+        .build();
+
+      @Override
+      public List<TransitMode> getModesOfStopLocation(StopLocation stop) {
+        if (stop.getGtfsVehicleType() != null) {
+          return List.of(stop.getGtfsVehicleType());
+        } else {
+          return List.copyOf(modes.get(stop));
+        }
+      }
+    };
     index = new LuceneIndex(graph, transitService);
+    mapper = new StopClusterMapper(transitService);
   }
 
   @Test
@@ -115,13 +148,13 @@ class LuceneIndexTest {
     @Test
     void stopClusters() {
       var result1 = index.queryStopClusters("alex").toList();
-      assertEquals(List.of(StopCluster.of(ALEXANDERPLATZ_STATION)), result1);
+      assertEquals(List.of(mapper.map(ALEXANDERPLATZ_STATION)), result1);
     }
 
     @Test
     void fuzzyStopClusters() {
       var result1 = index.queryStopClusters("arts").toList();
-      assertEquals(List.of(StopCluster.of(ARTS_CENTER).get()), result1);
+      assertEquals(List.of(mapper.map(ARTS_CENTER).get()), result1);
     }
 
     @Test
@@ -139,7 +172,23 @@ class LuceneIndexTest {
     )
     void stopClustersWithSpace(String query) {
       var result = index.queryStopClusters(query).toList();
-      assertEquals(List.of(StopCluster.of(FIVE_POINTS_STATION)), result);
+      assertEquals(List.of(mapper.map(FIVE_POINTS_STATION)), result);
+    }
+
+    @Test
+    void stopCode() {
+      var result = index.queryStopClusters(ARTS_CENTER.getCode()).toList();
+      assertEquals(1, result.size());
+      assertEquals(ARTS_CENTER.getName().toString(), result.get(0).name());
+    }
+
+    @Test
+    void modes() {
+      var result = index.queryStopClusters("westh").toList();
+      assertEquals(1, result.size());
+      var stop = result.get(0);
+      assertEquals(WESTHAFEN.getName().toString(), stop.name());
+      assertEquals(List.of(FERRY.name(), BUS.name()), stop.modes());
     }
   }
 }