+++ src/main/java/org/opentripplanner/astar/spi/AStarEdge.java
@@ -9,5 +9,5 @@ public interface AStarEdge<
 
   Vertex getToVertex();
 
-  State traverse(State u);
+  State[] traverse(State u);
 }