+++ src/main/java/org/opentripplanner/standalone/config/routerconfig/ServerConfig.java
@@ -34,6 +34,7 @@ domain, these are set tin the routing request.
        This does not include network latency nor waiting time in the HTTP server thread pool.
        The default value is `-1s` (no timeout).
        The timeout is applied to all APIs (REST, Transmodel , Legacy GraphQL).
+       The timeout is not enforced when the parallel routing OTP feature is in use.
         """
         )
         .asDuration(Duration.ofSeconds(-1));