+++ src/main/java/org/opentripplanner/street/model/edge/ElevatorHopEdge.java
@@ -81,7 +81,7 @@ public class ElevatorHopEdge extends Edge implements ElevatorEdge, WheelchairTra
   }
 
   @Override
-  public State traverse(State s0) {
+  public State[] traverse(State s0) {
     RoutingPreferences preferences = s0.getPreferences();
 
     StateEditor s1 = createEditorForDrivingOrWalking(s0, this);
@@ -91,7 +91,7 @@ public class ElevatorHopEdge extends Edge implements ElevatorEdge, WheelchairTra
         wheelchairAccessibility != Accessibility.POSSIBLE &&
         preferences.wheelchair().elevator().onlyConsiderAccessible()
       ) {
-        return null;
+        return State.empty();
       } else if (wheelchairAccessibility == Accessibility.NO_INFORMATION) {
         s1.incrementWeight(preferences.wheelchair().elevator().unknownCost());
       } else if (wheelchairAccessibility == Accessibility.NOT_POSSIBLE) {
@@ -102,15 +102,15 @@ public class ElevatorHopEdge extends Edge implements ElevatorEdge, WheelchairTra
     TraverseMode mode = s0.getNonTransitMode();
 
     if (mode == TraverseMode.WALK && !permission.allows(StreetTraversalPermission.PEDESTRIAN)) {
-      return null;
+      return State.empty();
     }
 
     if (mode == TraverseMode.BICYCLE && !permission.allows(StreetTraversalPermission.BICYCLE)) {
-      return null;
+      return State.empty();
     }
     // there are elevators which allow cars
     if (mode == TraverseMode.CAR && !permission.allows(StreetTraversalPermission.CAR)) {
-      return null;
+      return State.empty();
     }
 
     s1.incrementWeight(
@@ -123,7 +123,7 @@ public class ElevatorHopEdge extends Edge implements ElevatorEdge, WheelchairTra
         ? this.travelTime
         : (int) (preferences.street().elevator().hopTime() * this.levels)
     );
-    return s1.makeState();
+    return s1.makeStateArray();
   }
 
   @Override