+++ src/test/java/org/opentripplanner/routing/graph/SimpleConcreteEdge.java
@@ -19,14 +19,14 @@ public class SimpleConcreteEdge extends Edge {
   }
 
   @Override
-  public State traverse(State s0) {
+  public State[] traverse(State s0) {
     double d = getDistanceMeters();
     TraverseMode mode = s0.getNonTransitMode();
     int t = (int) (d / s0.getPreferences().getSpeed(mode, false));
     StateEditor s1 = s0.edit(this);
     s1.incrementTimeInSeconds(t);
     s1.incrementWeight(d);
-    return s1.makeState();
+    return s1.makeStateArray();
   }
 
   @Override