+++ src/test/java/org/opentripplanner/street/model/edge/StreetEdgeCostTest.java
@@ -182,6 +182,6 @@ class StreetEdgeCostTest {
     var state = new State(V1, request);
 
     assertEquals(0, state.weight);
-    return edge.traverse(state);
+    return edge.traverse(state)[0];
   }
 }