+++ docs/Changelog.md
@@ -14,6 +14,15 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Improve Graph updaters shutdown [#5092](https://github.com/opentripplanner/OpenTripPlanner/pull/5092)
 - Add deduplicated  to geocoding sandbox API [#5084](https://github.com/opentripplanner/OpenTripPlanner/pull/5084)
 - Fix flex timeshift [#5063](https://github.com/opentripplanner/OpenTripPlanner/pull/5063)
+- Merge norway traversal calculator into default [#5106](https://github.com/opentripplanner/OpenTripPlanner/pull/5106)
+- Use correct GTFS sequence number in vehicle position matcher [#5090](https://github.com/opentripplanner/OpenTripPlanner/pull/5090)
+- Fix SIRI ET PubSub updater shutdown [#5104](https://github.com/opentripplanner/OpenTripPlanner/pull/5104)
+- OSM: Break out of processing a malformed level map [#5096](https://github.com/opentripplanner/OpenTripPlanner/pull/5096)
+- Create unique SIRI-ET PubSub subscription [#5118](https://github.com/opentripplanner/OpenTripPlanner/pull/5118)
+- Handle JsonParseException [#5121](https://github.com/opentripplanner/OpenTripPlanner/pull/5121)
+- Add modes to geocoding  [#5115](https://github.com/opentripplanner/OpenTripPlanner/pull/5115)
+- Do not enforce API processing timeout for parallel routing [#5114](https://github.com/opentripplanner/OpenTripPlanner/pull/5114)
+- Rename bikeRental options to vehicleRental [#5089](https://github.com/opentripplanner/OpenTripPlanner/pull/5089)
 [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)